fix: migrate story to be nested under Split Button
